Changed country in ItemCountryEntity to One-One instead of Many-One (towards CountryEntity).
One pricing for an item only belong to one country.
This should be a uni-direction relationship, no need for the reverse to go from CountryEntity back to ItemCountryEntity.
If searching for items that are sold in a country, should use the bi-relationship from CountryEntity to ItemEntity first, then towards ItemCountryEntity.